% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_helpers.R
\name{zscore_numeric}
\alias{zscore_numeric}
\title{Center and Scale a Numeric Vector}
\usage{
zscore_numeric(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector with mean approximately 0 and standard deviation 1.
}
\description{
Standardizes a numeric vector to z-scores.
}
\examples{
zscore_numeric(c(10, 20, 30))
}
